Fix positioning, MVP application and vertex shader model matrix.
There were a lot of things wrong, took me quite a while to figure this
out. First of all, the view projection matrix was not applied on the
model matrix when drawing the objects. Second, the model matrix was not
applied on the actual vertex positions and normals, so when lighting
still considered their original values, leading to strange results.
Managed to update the vertex shader to take the model matrix into
account. Finally, fixed the transformations on the camera's and objects
such that the scene makes more sense. Still need to fix the
transformations in the weapons. Also need to clean up some code and
remove some temporary hacks (like in the fragment shader).